+++ FirebaseVertexAI/Tests/TestApp/Tests/Integration/IntegrationTests.swift
@@ -19,6 +19,7 @@ import FirebaseVertexAI
 import VertexAITestApp
 import XCTest
 
+// TODO(#14405): Migrate to Swift Testing and parameterize tests to run on both `v1` and `v1beta`.
 final class IntegrationTests: XCTestCase {
   // Set temperature, topP and topK to lowest allowed values to make responses more deterministic.
   let generationConfig = GenerationConfig(