+++ FirebaseVertexAI/CHANGELOG.md
@@ -5,6 +5,7 @@
   Note: This feature is in Public Preview, which means that the it is not
   subject to any SLA or deprecation policy and could change in
   backwards-incompatible ways.
+- [feature] Added support for modality-based token count. (#14406)
 
 # 11.6.0
 - [changed] The token counts from `GenerativeModel.countTokens(...)` now include